feat,wip(server): initially implement packet decoding of Login
I initially wrote an imperative code, but then decided to switch
to a more functional approach so it could be macro'd easier.

But apparently I couldn't break early from a `try_fold`.

I needed to loop over a (similarly) `Vec<(String, String)>`
to find the exact key, store them, and to return them.

Using the imperative style makes it easier, but since I was
visioning to create a macro, it seems like a lot that needed
to be changed on macros. So I decided to try out a functional
approach. I wanted to break early when all of the fields are
found, because who knows that places a large amount of keys in
their requests.

Or maybe I should just keep this and fix it later. It's that
one small perf nudge, that might not feel like much but
it's just so annoying to get optimized.